refactor(*): refactor utility imports to use new utility modules
